Fix QDoubleSpinBox slots

+++ Code/src/lutcustomize.h
@@ -32,8 +32,8 @@ private slots:
     void drawLine(double from, double to);
     void setRange();
     void on_okPushButton_clicked();
-    void on_fromSpinBox_valueChanged(int arg1);
-    void on_toSpinBox_valueChanged(int arg1);
+    void on_fromSpinBox_valueChanged(double arg1);
+    void on_toSpinBox_valueChanged(double arg1);
 
     void on_resetMinPushButton_clicked();
